Zug- und Flugzeugwege werden zurückgegeben

+++ src/main/java/travelcompare/restapi/provider/model/Route.java
@@ -24,7 +24,7 @@ public class Route {
 
     @Getter
     @Setter
-    private long distance;
+    private double distance;
 
     @Getter
     @Setter
